[CSPM] e2e test - move kubernetes helper to common lib
