+++ test/e2e/cws-tests/tests/lib/kubernetes.py
